refactor(data): split artistData.ts — extract JSON-LD schema to artist.schema.ts
artistData.ts: 1091 → 419 lines (ARTIST object + helpers only)
artist.schema.ts: new, 677 lines (DISAMBIGUATING_DESCRIPTION, ARTIST_SCHEMA_SAME_AS,
ARTIST_SCHEMA_BASE, Release types, DISCOGRAPHY, ARTIST_BUSINESS_SCHEMA, MUSICGROUP_SCHEMA)

7 consumer files updated to import schema exports from artist.schema.ts instead
of artistData.ts. No behavioral changes — pure mechanical extraction.

Zero test regressions (56 passing).
