+++ src/pages/MediaPage.tsx
@@ -2,7 +2,8 @@ import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { motion } from 'framer-motion';
 import { Newspaper, ExternalLink, Download, Image as ImageIcon, ShieldCheck } from 'lucide-react';
-import { ARTIST, ARTIST_SCHEMA_BASE, MUSICGROUP_SCHEMA } from '../data/artistData';
+import { ARTIST } from '../data/artistData';
+import { ARTIST_SCHEMA_BASE, MUSICGROUP_SCHEMA } from '../data/artist.schema';
 import { PUBLISHED_WORKS } from '../data/publishedWorks';
 import { useBranding } from '../contexts/BrandingContext';
 import { HeadlessSEO } from '../components/HeadlessSEO';
